chore(web): split project page tabs into individual components
